backend/pyproject.toml
[project]
name = "cohezy-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.0",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "SQLAlchemy==2.0.35",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.2.0",
  "redis==5.0.7",
  "python-multipart==0.0.9",
  "qrcode==7.4.2",
  "Pillow==10.4.0",
  "httpx==0.27.2",
  "orjson==3.10.7"
]

[project.optional-dependencies]
dev = ["pytest==8.3.2", "pytest-asyncio==0.23.8", "pytest-cov==5.0.0", "ruff==0.6.4", "black==24.8.0", "mypy==1.11.1"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
